Fix flaky whole-realm parse discovery test via indexer-settled seeding
The "clean realm returns status: passed" test in
run-parse-in-memory.spec.ts seeded its module + example + Spec with
per-file `client.write` + `waitForFile`. `waitForFile` only confirms the
source landed on disk (a source GET returns 200), not that the card is
indexed and searchable. Whole-realm parse discovers JSON examples
through `_federated-search` on Spec cards, so when the Spec's index — or
its resolved `linkedExamples` link — lagged the write, discovery returned
no example and it was dropped from `parseableFiles`, failing the
`toContain('ParseTestCard/example-1.json')` assertion intermittently.

Seed the three files in one `client.sync(..., { waitForIndex: true })`
batch instead — the same `_atomic` indexer-settled boundary the
run-instantiate-in-memory test already uses (and which does not flake).
Extract that helper (`seedFilesAndWaitForIndex`) from
instantiate-test-fixtures into a shared tests/helpers module so both
specs share one implementation.

Also log the shape of each JSON-example discovery at debug level, so the
agent-facing `run_parse` path — which can't use `waitForIndex` and can
still race a just-written Spec — makes a dropped example diagnosable.
