+++ packages/software-factory/tests/helpers/instantiate-test-fixtures.ts
@@ -5,12 +5,12 @@
  * tool). Keeping the card modules, examples, and specs in one place ensures
  * the two surfaces are exercised against identical inputs.
  */
-import { mkdirSync, mkdtempSync, rmSync, writeFileSync } from 'node:fs';
-import { tmpdir } from 'node:os';
-import { dirname, isAbsolute, join } from 'node:path';
+import { mkdirSync, writeFileSync } from 'node:fs';
+import { dirname, join } from 'node:path';
 
 import type { BoxelCLIClient } from '@cardstack/boxel-cli/api';
-import { expect } from '@playwright/test';
+
+import { seedFilesAndWaitForIndex } from './seed-and-wait-for-index.ts';
 
 // ---------------------------------------------------------------------------
 // Card modules
@@ -183,95 +183,6 @@ export function tagsCardSpecJson(): string {
 // ---------------------------------------------------------------------------
 // Seeding helpers
 // ---------------------------------------------------------------------------
-
-/**
- * Stage the given files in a fresh temp dir and push them to the realm
- * with `client.sync(..., { preferLocal: true, waitForIndex: true })`.
- * The `_atomic` upload appends `?waitForIndex=true`, so the realm-server
- * returns only after the indexer has processed every uploaded file.
- *
- * Why this shape instead of per-file `client.write` + a search-poll
- * gate: realm-side source POST indexing is async, so writing files
- * one-by-one and waiting for the realm to ack with GET (or for the
- * downstream `_federated-search` to surface them) is a polling race —
- * which is exactly the flake this skill was filed to address. CI
- * evidence on PR #4782 (run 25768256725) showed even a 90s poll
- * sometimes never sees the Spec surface in search. The `_atomic`
- * waitForIndex query param is the realm-server's first-class hook for
- * read-after-write consistency in tests; using it here trades a
- * one-shot push latency for a deterministic "indexer is settled"
- * boundary.
- */
-async function seedFilesAndWaitForIndex(
-  client: BoxelCLIClient,
-  realmUrl: string,
-  files: { path: string; content: string }[],
-): Promise<void> {
-  for (let { path } of files) {
-    if (isAbsolute(path) || path.split('/').includes('..')) {
-      throw new Error(
-        `seedFilesAndWaitForIndex path must be a realm-relative path under the staging dir; got ${JSON.stringify(path)}`,
-      );
-    }
-  }
-
-  // Retry the sync once on failure. The realm-server's /_atomic
-  // endpoint can return 500 ("Write Error") for transient causes —
-  // for example a concurrent indexing pass or a worker-side
-  // fileSerialization that doesn't recur on retry — and a one-shot
-  // retry with a short backoff lets the helper recover instead of
-  // surfacing the failure as a confusing test flake.
-  //
-  // Each attempt uses a fresh staging dir: `RealmSyncer` writes
-  // `.boxel-sync.json` even when the batch failed, recording hashes
-  // for the staged files. A retry against the same dir would see
-  // those hashes match the (unchanged) local files, classify the
-  // entries as "unchanged locally / deleted remotely" (since the
-  // failed batch never actually wrote them), and with `preferLocal:
-  // true` resolve as a noop — silently skipping the upload while
-  // reporting success. Fresh dir = no stale manifest = honest
-  // re-upload on attempt 2.
-  const maxAttempts = 2;
-  let syncResult: Awaited<ReturnType<typeof client.sync>> | undefined;
-  for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-    let stagingDir = mkdtempSync(join(tmpdir(), 'sf-instantiate-seed-'));
-    try {
-      for (let { path, content } of files) {
-        let absolute = join(stagingDir, path);
-        mkdirSync(dirname(absolute), { recursive: true });
-        writeFileSync(absolute, content);
-      }
-      syncResult = await client.sync(realmUrl, stagingDir, {
-        preferLocal: true,
-        waitForIndex: true,
-      });
-    } finally {
-      rmSync(stagingDir, { recursive: true, force: true });
-    }
-    if (!syncResult.hasError) {
-      if (attempt > 1) {
-        console.log(
-          `seedFilesAndWaitForIndex: sync succeeded on attempt ${attempt}/${maxAttempts} for ${realmUrl}`,
-        );
-      }
-      break;
-    }
-    console.log(
-      `seedFilesAndWaitForIndex: sync attempt ${attempt}/${maxAttempts} for ${realmUrl} failed: ${
-        syncResult.error ?? '(no error message)'
-      }`,
-    );
-    if (attempt < maxAttempts) {
-      await new Promise((resolve) => setTimeout(resolve, 500 * attempt));
-    }
-  }
-  expect(
-    syncResult!.hasError,
-    `seed sync to ${realmUrl} reported an error after ${maxAttempts} attempt(s): ${
-      syncResult!.error ?? '(no error message)'
-    }`,
-  ).toBe(false);
-}
 
 /**
  * Seed `instantiate-test-card.gts` + one linked example + a Spec that
